+++ gulpfile.js/tasks/inject-component-css.js
@@ -1,27 +1,39 @@
 // @formatter:off
 
-var requireCached = require('../src/gulp/require-cached');
-var config = require('../config');
-var runSequence = require('run-sequence');
-var path = require('path');
+const requireCached = require('../src/gulp/require-cached');
+const config = require('../config');
+const runSequence = require('run-sequence');
+const path = require('path');
+const walkFileListSync = require('../src/node/file/walk-file-list-sync');
 
-var gulp = requireCached('gulp');
-var inject = requireCached('gulp-inject');
-var sort = requireCached('gulp-sort');
+const gulp = requireCached('gulp');
+const inject = requireCached('gulp-inject');
 
 // @formatter:on
+
+const createComponentsArray = (folder) => {
+
+    const components = walkFileListSync(config.source.getPath(folder), 'stylesheet');
+    return [].reduce.call(components, (data, component) => {
+
+        data.push(path.join(component, '*.scss'));
+        return data;
+
+    }, []);
+
+}
 
 /**
  * Task for injecting scss
  */
 gulp.task('inject-component-css', function () {
 
-    var injectComponentsFiles = gulp.src([
-        path.join(config.source.getPath('components'), '**/*.scss')
-    ], { read: false });
+    const components = createComponentsArray('components');
+
+    var injectComponentsFiles = gulp.src(components, { read: false });
 
     var injectComponentsOptions = {
-        transform: (filePath) => `@import '${filePath}';`,
+        transform: (filePath) => `@import '${filePath.replace('.scss', '')}';`,
         starttag: '/* components:scss */',
         endtag: '/* endinject */',
         addRootSlash: false