+++ source/javascript/main.js
@@ -3,18 +3,17 @@
 /*------------------------------------*\
  * JS Main entry file
  \*------------------------------------*/
-import 'babel-polyfill';
 
 // Import utilities.
 import env from './src/system/environment';
-import moduleInit from './src/modules/util/module-init';
-import { initServiceWorker, removeServiceWorker } from './src/modules/util/sw';
+import moduleInit from 'utilities/module-init';
+import { initServiceWorker, removeServiceWorker } from 'utilities/sw';
 
 // Import Singletons
-import './src/modules/util/detect-touch';
+import 'utilities/detect-touch';
 
 // Import modules.
-import ExampleModule from './src/modules/example';
+import ExampleModule from 'components/example';
 
 if (!env.isLocal()) {
     initServiceWorker();